Admin Control

+++ src/main/java/net/engineeringdigest/journalApp/service/JournalEntryService.java
@@ -30,7 +30,7 @@ public class JournalEntryService {
             journalEntry.setDate(LocalDateTime.now());
             JournalEntry saved = journalEntryRepository.save(journalEntry);
             user.getJournalEntries().add(saved);
-            userService.saveEntry(user);
+            userService.saveUser(user);
         }
         catch (Exception e) {
         }
@@ -54,10 +54,21 @@ public class JournalEntryService {
         return journalEntryRepository.findById(id);
     }
 
-    public void deleteById(ObjectId id, String userName) {
-        User user =userService.findByUserName(userName);
-        user.getJournalEntries().removeIf(x -> x.getId().equals(id));
-        userService.saveEntry(user);
-        journalEntryRepository.deleteById(id);
+    @Transactional
+    public boolean deleteById(ObjectId id, String userName) {
+        boolean removed;
+        try {
+            User user =userService.findByUserName(userName);
+            removed =  user.getJournalEntries().removeIf(x -> x.getId().equals(id));
+            if (removed) {
+                userService.saveUser(user);
+                journalEntryRepository.deleteById(id);
+            }
+        }
+       catch (Exception e) {
+            System.out.println(e);
+            throw new RuntimeException("An error has occurred while deleting the entry "+e);
+       }
+        return removed;
     }
 }
